requirements.txt
Flask>=2.3
requests>=2.31
google-genai>=0.3.0
